+++ misc/services/mobile-webui/mobile-webui-frontend/src/components/BarcodeScanner/BarcodeScannerFooter.jsx
@@ -1,50 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { MODE, PRIORITY } from './useBarcodeScannerModes';
+import { MODE } from './useBarcodeScannerModes';
 import ButtonWithIndicator from '../buttons/ButtonWithIndicator';
 
-// "Back to scanner" — when in MANUAL, the button targets the highest-priority scanner mode
-// that is currently enabled. HARDWARE wins over CAMERA (per PRIORITY in useBarcodeScannerModes).
-// If neither scanner mode is enabled the button is hidden — there's nowhere to go back to.
-const SCANNER_MODES = PRIORITY.filter((m) => m === MODE.HARDWARE || m === MODE.CAMERA);
-
-const BACK_TO_SCANNER_CAPTION_KEY = {
-  [MODE.HARDWARE]: 'components.BarcodeScannerComponent.useHardwareScanner',
-  [MODE.CAMERA]: 'components.BarcodeScannerComponent.scanWithCamera',
-};
-const BACK_TO_SCANNER_ICON = {
-  [MODE.HARDWARE]: 'fa-barcode',
-  [MODE.CAMERA]: 'fa-camera',
-};
-
 /*
- * Footer button cluster. Renders the mode-switching buttons applicable to the current state
- * and signals the parent via ONE callback: onModeSelected(targetMode).
+ * Footer button cluster — shown only in scanner modes (HARDWARE / CAMERA).
  *
- * Each button resolves its own target mode locally — the footer is the source of truth for
- * "which mode does this button switch to":
- *   - HW⇄CAM toggle  → the opposite scanner mode of the current activeMode
- *   - Back-to-scanner → first enabled of HARDWARE|CAMERA (PRIORITY order)
+ * In MANUAL mode the footer renders nothing (returns null) so the empty area below the
+ * manual input stays free for the mobile virtual keyboard. The "back to scanner" button
+ * lives inside ManualModePanel itself in that mode (stacked under Submit, anchored at the
+ * top of the panel) so it doesn't get overlapped by the keyboard.
+ *
+ * Two buttons can show here, both signal the parent via onModeSelected(targetMode):
+ *   - HW⇄CAM toggle  → opposite scanner mode of the current activeMode
  *   - Enter manually → MODE.MANUAL
- *
- * The parent's only job is to apply the new mode (typically `onModeSelected={setActiveMode}`).
  */
 const BarcodeScannerFooter = ({ activeMode, enabledModes, onModeSelected }) => {
-  const isManualEnabled = !!enabledModes?.manual;
-
-  // HW↔Camera toggle only when BOTH scanner modes are enabled AND the user is currently in a
-  // scanner mode. Hidden in MANUAL — the dedicated back-to-scanner button covers that case.
+  // Both buttons are gated on activeMode !== MANUAL → in MANUAL the footer is empty and
+  // the early return below kicks in.
   const isHardwareCameraToggleShown = !!(enabledModes?.hardware && enabledModes?.camera) && activeMode !== MODE.MANUAL;
+  const isManualButtonShown = !!enabledModes?.manual && activeMode !== MODE.MANUAL;
 
-  // Manual button — only useful when not already in manual mode.
-  const isManualButtonShown = isManualEnabled && activeMode !== MODE.MANUAL;
-
-  // Back-to-scanner — only relevant from MANUAL; resolves to the highest-priority enabled
-  // scanner mode (HARDWARE > CAMERA). Undefined if no scanner mode is enabled → button hidden.
-  const backToScannerTarget = activeMode === MODE.MANUAL ? SCANNER_MODES.find((m) => enabledModes?.[m]) : undefined;
-  const isBackToScannerShown = !!backToScannerTarget;
-
-  if (!isManualButtonShown && !isHardwareCameraToggleShown && !isBackToScannerShown) {
+  if (!isManualButtonShown && !isHardwareCameraToggleShown) {
     return null;
   }
 
@@ -64,15 +41,6 @@ const BarcodeScannerFooter = ({ activeMode, enabledModes, onModeSelected }) => {
           additionalCssClass="barcode-scanner-btn"
           onClick={() => onModeSelected(toggleTarget)}
           testId="barcode-scanner-toggle-hw-camera"
-        />
-      )}
-      {isBackToScannerShown && (
-        <ButtonWithIndicator
-          captionKey={BACK_TO_SCANNER_CAPTION_KEY[backToScannerTarget]}
-          typeFASIconName={BACK_TO_SCANNER_ICON[backToScannerTarget]}
-          additionalCssClass="barcode-scanner-btn"
-          onClick={() => onModeSelected(backToScannerTarget)}
-          testId="barcode-scanner-back-to-scanner"
         />
       )}
       {isManualButtonShown && (
